[JIT] Enable @ignore and @unused syntax for ignoring properties
**Summary**
This commit enables `@ignore` and `@unused` syntax for ignoring
properties, but works only if these decorators are applied to a function
before the @property decorator. Despite this, ignoring properties is
more intuitive with this feature enabled.

**Test Plan**
This commit updates the existing unit tests for class type and module
properties to test properties ignored using `@ignore` and `@unused`.
Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 7a545729ca5efff280658a00e32e5053d7619f9d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45261